Moved ReconContext, changed get_optimization_parameters in dataset_models.py
